AddTransient<UsersService>

+++ OnlineBlog.Server/Controllers/AccountController.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
-using OnlineBlog.Server.Data;
 using OnlineBlog.Server.Models;
 using OnlineBlog.Server.Services;
 using System.IdentityModel.Tokens.Jwt;
@@ -13,45 +12,43 @@ namespace OnlineBlog.Server.Controllers
     /// Контроллер аккаунтов (создать/удалить/изменить/обновить пользователя)
     /// </summary>
     [ApiController]
-    [Authorize]
-    [Route("[controller]")]    
+    [Authorize] // доступ только авторизованным
+    [Route("[controller]")] // маршрут до контроллеров   
     public class AccountController : ControllerBase
     {
         private UsersService _usersService;
-        public AccountController(AppDataContext dataContext)
+        public AccountController(UsersService usersService)
         {
-            _usersService = new UsersService(dataContext);
+            _usersService = usersService;
         }
 
+        #region CRUD
         /// <summary>
         /// Получить пользователя
         /// </summary>
-        [HttpGet]        
+        [HttpGet]
         public IActionResult Get()
         {
             var currentUserEmail = HttpContext.User.Identity.Name;
-            var currentUser = _usersService.GetUserByLogin(currentUserEmail);
-            if (currentUser == null)
-            {
-                return NotFound();
-            }
-
-            return Ok(new UserModel()
-            {
-                Id = currentUser.Id,
-                Email = currentUser.Email,
-                FirstName = currentUser.FirstName,
-                LastName = currentUser.LastName,
-                Description = currentUser.Description,
-                Photo = currentUser.Photo
-            });
+            var currentUser = _usersService.GetUserByEmail(currentUserEmail);
+            return currentUser == null
+                ? NotFound()
+                : Ok(new UserModel()
+                {
+                    Id = currentUser.Id,
+                    Email = currentUser.Email,
+                    FirstName = currentUser.FirstName,
+                    LastName = currentUser.LastName,
+                    Description = currentUser.Description,
+                    Photo = currentUser.Photo
+                });
         }
 
         /// <summary>
         /// Создать пользователя
         /// </summary>
         [HttpPost]
-        public ActionResult<UserModel> Create([FromBody] UserModel user)
+        public ActionResult<UserModel> Create(UserModel user)
         {
             var newUser = _usersService.Create(user);
             return Ok(newUser);
@@ -64,10 +61,10 @@ namespace OnlineBlog.Server.Controllers
         public ActionResult<UserModel> Update(UserModel user)
         {
             var currentUserEmail = HttpContext.User.Identity.Name;
-            var currentUser = _usersService.GetUserByLogin(currentUserEmail);
+            var currentUser = _usersService.GetUserByEmail(currentUserEmail);
             if (currentUser == null)
             {
-                return BadRequest();
+                return NotFound();
             }
             _usersService.Update(user);
             return Ok(user);
@@ -80,11 +77,17 @@ namespace OnlineBlog.Server.Controllers
         public IActionResult Delete()
         {
             var currentUserEmail = HttpContext.User.Identity.Name;
-            var currentUser = _usersService.GetUserByLogin(currentUserEmail);
+            var currentUser = _usersService.GetUserByEmail(currentUserEmail);
+            if (currentUser == null)
+            {
+                return NotFound();
+            }
             _usersService.Delete(currentUser);
             return Ok();
         }
+        #endregion
 
+        #region Token
         /// <summary>
         /// Получить jwt токен
         /// </summary>
@@ -115,7 +118,7 @@ namespace OnlineBlog.Server.Controllers
 
             var encodedJwt = new JwtSecurityTokenHandler().WriteToken(jwt);
 
-            //return token
+            // return token
             var tokenModel = new AuthToken(
                 minutes: AuthOptions.LIFETIME,
                 accessToken: encodedJwt,
@@ -123,5 +126,6 @@ namespace OnlineBlog.Server.Controllers
                 userId: identity.Value.id);
             return Ok(tokenModel);
         }
+        #endregion
     }
 }
